+++ include/a3d/scene/Scene.h
@@ -42,7 +42,7 @@ namespace a3d {
 
 	public:
 
-		using UpdateCallback =				std::function<void(Scene& scene, float time, float deltaTime)>;
+		using UpdateCallback =				std::function<void(Scene& scene, double time, double deltaTime)>;
 
 /*********************************************************************************************
 	Public Static Member Functions
