+++ tests/005-materials-lights/main.cc
@@ -34,9 +34,9 @@ constexpr bool 					ORTHO_CAMERA =			false;
 constexpr float					MOUSE_SENSITIVITY =		0.5;
 
 
-void UpdateCallback(Scene& scene, float time, float deltaTime);
-void WillRenderCallback(VisualWorld& world, float time, float deltaTime);
-void DidRenderCallback(VisualWorld& world, float time, float deltaTime);
+void UpdateCallback(Scene& scene, double time, double deltaTime);
+void WillRenderCallback(VisualWorld& world, double time, double deltaTime);
+void DidRenderCallback(VisualWorld& world, double time, double deltaTime);
 
 
 void InitLog();
@@ -205,7 +205,7 @@ int main(int argc, const char* argv[]) {
 	Scene Callbacks
  ***************************************************************************************/
 
-void UpdateCallback(Scene& scene, float time, float deltaTime) {
+void UpdateCallback(Scene& scene, double time, double deltaTime) {
 	static int invocations = 0;
 	if (invocations == 2) {
 		double time = utils::Time() - g_startTime;
@@ -345,20 +345,20 @@ void UpdateCallback(Scene& scene, float time, float deltaTime) {
 			}
 
 			if(keysDown.count(Key::W)) {
-				vec3 positionDelta = deltaTime * MOVE_SPEED * camForward;
+				vec3 positionDelta = (float)deltaTime * MOVE_SPEED * camForward;
 				pov->position(pov->position() + positionDelta);
 			}
 			else if(keysDown.count(Key::S)) {
-				vec3 positionDelta = deltaTime * MOVE_SPEED * -camForward;
+				vec3 positionDelta = (float)deltaTime * MOVE_SPEED * -camForward;
 				pov->position(pov->position() + positionDelta);
 			}
 
 			if(keysDown.count(Key::A)) {
-				vec3 positionDelta = deltaTime * MOVE_SPEED * -camRight;
+				vec3 positionDelta = (float)deltaTime * MOVE_SPEED * -camRight;
 				pov->position(pov->position() + positionDelta);
 			}
 			else if(keysDown.count(Key::D)) {
-				vec3 positionDelta = deltaTime * MOVE_SPEED * camRight;
+				vec3 positionDelta = (float)deltaTime * MOVE_SPEED * camRight;
 				pov->position(pov->position() + positionDelta);
 			}
 
@@ -367,7 +367,7 @@ void UpdateCallback(Scene& scene, float time, float deltaTime) {
 				if (keysDown.count(Key::LeftShift)) {
 					direction = -1;
 				}
-				vec3 positionDelta = deltaTime * MOVE_SPEED * moveMultiplier * camUp;
+				vec3 positionDelta = (float)deltaTime * MOVE_SPEED * moveMultiplier * camUp;
 				pov->position(pov->position() + positionDelta * direction);
 			}
 		}
@@ -399,11 +399,11 @@ void UpdateCallback(Scene& scene, float time, float deltaTime) {
 	VisualWorld Callbacks
  ***************************************************************************************/
 
-void WillRenderCallback(VisualWorld& world, float time, float deltaTime) {
+void WillRenderCallback(VisualWorld& world, double time, double deltaTime) {
 	LOG_T(g_logger, "world: {:p}, time: {}, deltaTime: {}", (void*)&world, time, deltaTime);
 }
 
-void DidRenderCallback(VisualWorld& world, float time, float deltaTime) {
+void DidRenderCallback(VisualWorld& world, double time, double deltaTime) {
 	LOG_T(g_logger, "world: {:p}, time: {}, deltaTime: {}", (void*)&world, time, deltaTime);
 }
